pyproject.toml
[project]
name = "commercial-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
]
[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pandas>=2.3.2",
    "pre-commit>=4.3.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.24.0",
    "aiosqlite>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
